refactor: 메뉴 관련 기능 의존성 변경

+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
-    int countByIdIn(List<Long> ids);
+    long countByIdIn(List<Long> ids);
 }
